+++ alignment_computation.hh
@@ -9,21 +9,21 @@ void compute_ibm1_viterbi_alignment(const Storage1D<uint>& source_sentence,
                                     const Math2D::Matrix<uint>& slookup,
                                     const Storage1D<uint>& target_sentence,
                                     const SingleWordDictionary& dict,
-                                    Storage1D<uint>& viterbi_alignment);
+                                    Storage1D<ushort>& viterbi_alignment);
 
 void compute_ibm2_viterbi_alignment(const Storage1D<uint>& source_sentence,
                                     const Math2D::Matrix<uint>& slookup,
                                     const Storage1D<uint>& target_sentence,
                                     const SingleWordDictionary& dict,
                                     const Math2D::Matrix<double>& align_prob,
-                                    Storage1D<uint>& viterbi_alignment);
+                                    Storage1D<ushort>& viterbi_alignment);
 
 void compute_fullhmm_viterbi_alignment(const Storage1D<uint>& source_sentence,
                                        const Math2D::Matrix<uint>& slookup,
                                        const Storage1D<uint>& target_sentence,
                                        const SingleWordDictionary& dict,
                                        const Math2D::Matrix<double>& align_prob,
-                                       Storage1D<uint>& viterbi_alignment);
+                                       Storage1D<ushort>& viterbi_alignment);
 
 double compute_ehmm_viterbi_alignment(const Storage1D<uint>& source_sentence,
                                       const Math2D::Matrix<uint>& slookup,
@@ -31,7 +31,7 @@ double compute_ehmm_viterbi_alignment(const Storage1D<uint>& source_sentence,
                                       const SingleWordDictionary& dict,
                                       const Math2D::Matrix<double>& align_prob,
                                       const Math1D::Vector<double>& initial_prob,
-                                      Storage1D<uint>& viterbi_alignment, 
+                                      Storage1D<ushort>& viterbi_alignment, 
 				      bool internal_mode = false, bool verbose = false);
 
 void compute_ehmm_optmarginal_alignment(const Storage1D<uint>& source_sentence,
@@ -40,7 +40,7 @@ void compute_ehmm_optmarginal_alignment(const Storage1D<uint>& source_sentence,
                                         const SingleWordDictionary& dict,
                                         const Math2D::Matrix<double>& align_prob,
                                         const Math1D::Vector<double>& initial_prob,
-                                        Storage1D<uint>& optmarginal_alignment);
+                                        Storage1D<ushort>& optmarginal_alignment);
 
 
 #endif
