+++ ibm1_training.hh
@@ -20,8 +20,8 @@ void train_ibm1(const Storage1D<Storage1D<uint> >& source,
                 uint nSourceWords, uint nTargetWords,
                 SingleWordDictionary& dict,
                 uint nIterations,
-                std::map<uint,std::set<std::pair<uint,uint> > >& sure_ref_alignments,
-                std::map<uint,std::set<std::pair<uint,uint> > >& possible_ref_alignments,
+                std::map<uint,std::set<std::pair<ushort,ushort> > >& sure_ref_alignments,
+                std::map<uint,std::set<std::pair<ushort,ushort> > >& possible_ref_alignments,
                 const floatSingleWordDictionary& prior_weight,
                 bool smoothed_l0 = false, double l0_beta = 1.0);
 
@@ -32,8 +32,8 @@ void train_ibm1_gd_stepcontrol(const Storage1D<Storage1D<uint> >& source,
                                uint nSourceWords, uint nTargetWords,
                                SingleWordDictionary& dict,
                                uint nIterations,
-                               std::map<uint,std::set<std::pair<uint,uint> > >& sure_ref_alignments,
-                               std::map<uint,std::set<std::pair<uint,uint> > >& possible_ref_alignments,
+                               std::map<uint,std::set<std::pair<ushort,ushort> > >& sure_ref_alignments,
+                               std::map<uint,std::set<std::pair<ushort,ushort> > >& possible_ref_alignments,
                                const floatSingleWordDictionary& prior_weight,
                                bool smoothed_l0 = false, double l0_beta = 1.0);
 
@@ -51,8 +51,8 @@ void ibm1_viterbi_training(const Storage1D<Storage1D<uint> >& source,
                            uint nSourceWords, uint nTargetWords,
                            SingleWordDictionary& dict,
                            uint nIterations,
-                           std::map<uint,std::set<std::pair<uint,uint> > >& sure_ref_alignments,
-                           std::map<uint,std::set<std::pair<uint,uint> > >& possible_ref_alignments,
+                           std::map<uint,std::set<std::pair<ushort,ushort> > >& sure_ref_alignments,
+                           std::map<uint,std::set<std::pair<ushort,ushort> > >& possible_ref_alignments,
                            const floatSingleWordDictionary& prior_weight);
 
 double prob_penalty(double x, double beta);
